+++ FunctionTest/src/main.cpp
@@ -12,7 +12,7 @@ int main()
         std::cout<<"loop"<<i<<"\n";
         std::cout<<"Pos: "<<LAnk1.ReadPos()<<std::endl;
 
-        std::array<int,8> adcRes = adc1.ReadData();
+        std::array<u_int16_t,8> adcRes = adc1.ReadData();
         std::cout<<"adc: ";
         for(unsigned i2 =0;i2<adcRes.size();i2++){
             std::cout<<adcRes[i2]<<',';
